requirements.txt
dash
dash-bootstrap-components
Flask
gunicorn
matplotlib
numpy
opencv-python
pandas
Pillow
plotly
pytesseract
pytest
python-dotenv
requests

